Finished off compiler and working on file splitting now
